+++ packages/client/src/Call.ts
@@ -157,19 +157,22 @@ export class Call {
   readonly camera: CameraManager;
 
   /**
-   * Device manager for the microhpone
+   * Device manager for the microphone.
    */
   readonly microphone: MicrophoneManager;
+
+  /**
+   * Device manager for the speaker.
+   */
+  readonly speaker: SpeakerManager;
 
   /**
    * The DynascaleManager instance.
    */
   readonly dynascaleManager = new DynascaleManager(this);
 
-  /*
-   * Device manager for the speaker
-   */
-  readonly speaker: SpeakerManager;
+  subscriber?: Subscriber;
+  publisher?: Publisher;
 
   /**
    * Flag telling whether this call is a "ringing" call.
@@ -188,8 +191,6 @@ export class Call {
    */
   private readonly dispatcher = new Dispatcher();
 
-  private subscriber?: Subscriber;
-  private publisher?: Publisher;
   private trackSubscriptionsSubject = new BehaviorSubject<{
     type: DebounceType;
     data: TrackSubscriptionDetails[];
@@ -277,37 +278,6 @@ export class Call {
 
     this.camera = new CameraManager(this);
     this.microphone = new MicrophoneManager(this);
-
-    this.state.localParticipant$.subscribe(async (p) => {
-      // Mute via device manager
-      // If integrator doesn't use device manager, we mute using stopPublish
-      if (
-        !p?.publishedTracks.includes(TrackType.VIDEO) &&
-        this.publisher?.isPublishing(TrackType.VIDEO)
-      ) {
-        this.logger(
-          'info',
-          `Local participant's video track is muted remotely`,
-        );
-        await this.camera.disable();
-        if (this.publisher.isPublishing(TrackType.VIDEO)) {
-          this.stopPublish(TrackType.VIDEO);
-        }
-      }
-      if (
-        !p?.publishedTracks.includes(TrackType.AUDIO) &&
-        this.publisher?.isPublishing(TrackType.AUDIO)
-      ) {
-        this.logger(
-          'info',
-          `Local participant's audio track is muted remotely`,
-        );
-        await this.microphone.disable();
-        if (this.publisher.isPublishing(TrackType.AUDIO)) {
-          this.stopPublish(TrackType.AUDIO);
-        }
-      }
-    });
     this.speaker = new SpeakerManager();
   }
 
@@ -392,7 +362,7 @@ export class Call {
     this.leaveCallHooks.add(
       // handles the case when the user is blocked by the call owner.
       createSubscription(this.state.blockedUserIds$, async (blockedUserIds) => {
-        if (!blockedUserIds) return;
+        if (!blockedUserIds || blockedUserIds.length === 0) return;
         const currentUserId = this.currentUserId;
         if (currentUserId && blockedUserIds.includes(currentUserId)) {
           this.logger('info', 'Leaving call because of being blocked');
@@ -928,7 +898,6 @@ export class Call {
         connectionConfig,
         isDtxEnabled,
         isRedEnabled,
-        preferredVideoCodec: this.streamClient.options.preferredVideoCodec,
       });
     }
 
@@ -1898,10 +1867,10 @@ export class Call {
       this.microphone.state.mediaStream &&
       !this.publisher?.isPublishing(TrackType.AUDIO)
     ) {
-      this.publishAudioStream(this.microphone.state.mediaStream);
+      await this.publishAudioStream(this.microphone.state.mediaStream);
     }
 
-    // Start mic if backend config speicifies, and there is no local setting
+    // Start mic if backend config specifies, and there is no local setting
     if (
       this.microphone.state.status === undefined &&
       this.state.settings?.audio.mic_default_on