+++ packages/client/src/events/index.ts
@@ -1,5 +1,6 @@
 export * from './call';
 export * from './call-permissions';
 export * from './internal';
+export * from './mutes';
 export * from './participant';
 export * from './speaker';