+++ packages/react-native-sdk/CHANGELOG.md
@@ -2,6 +2,39 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.0.13](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.12...@stream-io/video-react-native-sdk-0.0.13) (2023-09-13)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.1.0`
+* `@stream-io/video-react-bindings` updated to version `0.2.17`
+### [0.0.12](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.11...@stream-io/video-react-native-sdk-0.0.12) (2023-09-12)
+
+
+### Bug Fixes
+
+* **react-native:** remove  method to inform SDK about native permissions ([#1072](https://github.com/GetStream/stream-video-js/issues/1072)) ([53b4abd](https://github.com/GetStream/stream-video-js/commit/53b4abd80c17fc21d7f3a93dd103946ef0b3f080))
+
+### [0.0.11](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.10...@stream-io/video-react-native-sdk-0.0.11) (2023-09-12)
+
+
+### Bug Fixes
+
+* **react-native:** add missing commonjs support ([#1075](https://github.com/GetStream/stream-video-js/issues/1075)) ([c9e4e7d](https://github.com/GetStream/stream-video-js/commit/c9e4e7df73c8568286afe18ec7816aa69836b1c7))
+
+### [0.0.10](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.9...@stream-io/video-react-native-sdk-0.0.10) (2023-09-11)
+
+
+### Bug Fixes
+
+* **react-native:** add missing expo plugin in npm pack ([bfeb79d](https://github.com/GetStream/stream-video-js/commit/bfeb79d2a540f627e8fcefbe1524cf634307ce84))
+
+### [0.0.9](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.8...@stream-io/video-react-native-sdk-0.0.9) (2023-09-11)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.1.0`
+* `@stream-io/video-react-bindings` updated to version `0.2.16`
 ### [0.0.8](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-native-sdk-0.0.7...@stream-io/video-react-native-sdk-0.0.8) (2023-09-07)